refactor: make process code dir, move label, select, and split to it
