+++ src/readii/process/split.py
